Clean up repo: docs consolidation, Actions deploy, WebP images
- Replace husky build-and-commit-docs deploy with a GitHub Actions
  Pages workflow (deploy.yml); build output moves to dist/, gh-pages
  and husky dependencies removed
- Restore missing project screenshots and hero portrait as WebP
  (~3.7MB of PNGs -> ~320KB), fixing unresolved imports on master
- Consolidate stale docs into SYSTEM_DESIGN.md (rewritten for v2),
  new CLAUDE.md, and an updated README
- Delete CRA-era public/index.html and unused og-image.png
